Update fetch policy for current User
It is not possible to fetch the cart and products on a user during sign in.
An anonymous client is not allowed to do that. It is thus not possible to correctly set the cart and product fields in Apollo cache for the User/authenticatedItem.

We need to refetch the user from the network as soon as we are signed in.